RavenDB-18212 moving processors around
