+++ src/Raven.Server/Documents/Handlers/Processors/SampleData/AbstractSampleDataHandlerProcessorForGetSampleDataClasses.cs
@@ -4,7 +4,7 @@ using Raven.Server.Web;
 using Raven.Server.Web.Studio;
 using Sparrow.Json;
 
-namespace Raven.Server.Documents.Handlers.Processors
+namespace Raven.Server.Documents.Handlers.Processors.SampleData
 {
     internal abstract class AbstractSampleDataHandlerProcessorForGetSampleDataClasses<TRequestHandler, TOperationContext> : AbstractHandlerProcessor<TRequestHandler, TOperationContext>
         where TRequestHandler : RequestHandler
