+++ src/Raven.Server/Documents/Handlers/Processors/SampleData/SampleDataHandlerProcessorForGetSampleDataClasses.cs
@@ -1,7 +1,7 @@
 ﻿using JetBrains.Annotations;
 using Raven.Server.ServerWide.Context;
 
-namespace Raven.Server.Documents.Handlers.Processors
+namespace Raven.Server.Documents.Handlers.Processors.SampleData
 {
     internal class SampleDataHandlerProcessorForGetSampleDataClasses : AbstractSampleDataHandlerProcessorForGetSampleDataClasses<DatabaseRequestHandler, DocumentsOperationContext>
     {
