+++ src/Raven.Server/Documents/Handlers/Processors/Stats/AbstractStatsHandlerProcessorForGetDetailedDatabaseStatistics.cs
@@ -6,7 +6,7 @@ using Raven.Server.Json;
 using Raven.Server.Web;
 using Sparrow.Json;
 
-namespace Raven.Server.Documents.Handlers.Processors
+namespace Raven.Server.Documents.Handlers.Processors.Stats
 {
     internal abstract class AbstractStatsHandlerProcessorForGetDetailedDatabaseStatistics<TRequestHandler, TOperationContext> : AbstractHandlerProxyReadProcessor<DetailedDatabaseStatistics, TRequestHandler, TOperationContext>
         where TRequestHandler : RequestHandler
