+++ src/Raven.Server/Documents/Handlers/Processors/Stats/StatsHandlerProcessorForGetDatabaseStatistics.cs
@@ -6,9 +6,8 @@ using Raven.Client.Documents.Operations;
 using Raven.Client.Http;
 using Raven.Server.Documents.Indexes;
 using Raven.Server.ServerWide.Context;
-using Sparrow.Json;
 
-namespace Raven.Server.Documents.Handlers.Processors
+namespace Raven.Server.Documents.Handlers.Processors.Stats
 {
     internal class StatsHandlerProcessorForGetDatabaseStatistics : AbstractStatsHandlerProcessorForGetDatabaseStatistics<DatabaseRequestHandler, DocumentsOperationContext>
     {
