+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/SampleData/ShardedSampleDataHandlerProcessorForGetSampleDataClasses.cs
@@ -1,8 +1,8 @@
 ﻿using JetBrains.Annotations;
-using Raven.Server.Documents.Handlers.Processors;
+using Raven.Server.Documents.Handlers.Processors.SampleData;
 using Raven.Server.ServerWide.Context;
 
-namespace Raven.Server.Documents.Sharding.Handlers.Processors
+namespace Raven.Server.Documents.Sharding.Handlers.Processors.SampleData
 {
     internal class ShardedSampleDataHandlerProcessorForGetSampleDataClasses : AbstractSampleDataHandlerProcessorForGetSampleDataClasses<ShardedRequestHandler, TransactionOperationContext>
     {
