+++ src/Raven.Server/Documents/Sharding/Handlers/ShardedCompareExchangeHandler.cs
@@ -1,7 +1,9 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents.Sharding.Handlers.Processors;
+using Raven.Server.Documents.Sharding.Handlers.Processors.CompareExchange;
 using Raven.Server.Routing;
 using Raven.Server.Web.System.Processors;
+using Raven.Server.Web.System.Processors.CompareExchange;
 
 namespace Raven.Server.Documents.Sharding.Handlers;
