+++ src/Raven.Server/Documents/Sharding/Handlers/ShardedSampleDataHandler.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents.Sharding.Handlers.Processors;
+using Raven.Server.Documents.Sharding.Handlers.Processors.SampleData;
 using Raven.Server.Routing;
 
 namespace Raven.Server.Documents.Sharding.Handlers
