+++ src/Raven.Server/Documents/Sharding/Handlers/ShardedStudioStatsHandler.cs
@@ -1,6 +1,7 @@
 ﻿extern alias NGC;
 using System.Threading.Tasks;
 using Raven.Server.Documents.Sharding.Handlers.Processors;
+using Raven.Server.Documents.Sharding.Handlers.Processors.Studio;
 using Raven.Server.Routing;
 
 namespace Raven.Server.Documents.Sharding.Handlers
