+++ src/Raven.Server/Web/Studio/SampleDataHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Handlers.Processors;
+using Raven.Server.Documents.Handlers.Processors.SampleData;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide.Context;
