+++ src/Raven.Server/Web/Studio/StudioStatsHandler.cs
@@ -1,6 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Raven.Server.Documents;
 using Raven.Server.Documents.Handlers.Processors;
+using Raven.Server.Documents.Handlers.Processors.Studio;
 using Raven.Server.Routing;
 
 namespace Raven.Server.Web.Studio
