+++ src/Raven.Server/Web/System/CompareExchangeHandler.cs
@@ -2,6 +2,7 @@
 using Raven.Server.Documents;
 using Raven.Server.Routing;
 using Raven.Server.Web.System.Processors;
+using Raven.Server.Web.System.Processors.CompareExchange;
 
 namespace Raven.Server.Web.System
 {
