+++ src/Raven.Server/Web/System/Processors/CompareExchange/CompareExchangeHandlerProcessorForDeleteCompareExchangeValue.cs
@@ -8,7 +8,7 @@ using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 
-namespace Raven.Server.Web.System.Processors;
+namespace Raven.Server.Web.System.Processors.CompareExchange;
 
 internal class CompareExchangeHandlerProcessorForDeleteCompareExchangeValue : AbstractHandlerProcessor<RequestHandler, TransactionOperationContext>
 {
