+++ src/Raven.Server/Web/System/Processors/CompareExchange/CompareExchangeHandlerProcessorForPutCompareExchangeValue.cs
@@ -7,7 +7,7 @@ using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 
-namespace Raven.Server.Web.System.Processors;
+namespace Raven.Server.Web.System.Processors.CompareExchange;
 
 public class CompareExchangeHandlerProcessorForPutCompareExchangeValue : IDisposable
 {
